+++ addons/website_slides/__manifest__.py
@@ -19,11 +19,17 @@ Share and Publish Videos, Presentations and Documents'
 """,
     'depends': ['website', 'website_mail'],
     'data': [
+        'views/assets.xml',
         'views/res_config_settings_views.xml',
-        'views/website_slides.xml',
-        'views/website_slides_embed.xml',
-        'views/website_slides_backend.xml',
-        'data/website_slides_data.xml',
+        'views/website_slides_templates.xml',
+        'views/website_slides_embed_templates.xml',
+        'views/slide_slide_views.xml',
+        'views/slide_channel_views.xml',
+        'views/website_slides_menu_views.xml',
+        'data/website_slides_ir_data.xml',
+        'data/mail_data.xml',
+        'data/slide_data.xml',
+        'data/website_data.xml',
         'security/ir.model.access.csv',
         'security/website_slides_security.xml'
     ],
