+++ addons/website_slides/data/slide_data.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data noupdate="1">
+         <record model="slide.channel" id="channel_public">
+            <field name="name">Public Channel</field>
+            <field name="promote_strategy">most_viewed</field>
+            <field name="website_published" eval="True"/>
+            <field name="visibility">public</field>
+            <field name="description">Default channel for slides, all public users can access content of this channel.</field>
+        </record>
+        <record model="slide.channel" id="channel_partial">
+            <field name="name">Partner Channel</field>
+            <field name="website_published" eval="True"/>
+            <field name="group_ids" eval="[(4, ref('base.group_portal')), (4, ref('base.group_user'))]"/>
+            <field name="upload_group_ids" eval="[(4, ref('base.group_user'))]"/>
+            <field name="visibility">partial</field>
+            <field name="description">Used to publish slides in partner network privately.</field>
+        </record>
+        <record model="slide.channel" id="channel_private">
+            <field name="name">Employee Channel</field>
+            <field name="group_ids" eval="[(4, ref('base.group_user'))]"/>
+            <field name="visibility">private</field>
+            <field name="description">Used to publish internal slides of company.</field>
+        </record>
+    </data>
+</odoo>
