+++ addons/website_slides/data/website_slides_ir_data.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data noupdate="1">
+        <!-- jump to Channel on installation complete -->
+        <record id="action_open_channels" model="ir.actions.act_url">
+            <field name="name">Channels</field>
+            <field name="target">self</field>
+            <field name="url" eval="'/slides'"/>
+        </record>
+        <record id="base.open_menu" model="ir.actions.todo">
+            <field name="action_id" ref="action_open_channels"/>
+            <field name="state">open</field>
+        </record>
+    </data>
+</odoo>
