+++ addons/website_slides/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from . import slides
+from . import slide_slide
+from . import slide_channel
 from . import res_config_settings
 from . import website
