+++ addons/website_slides/views/slide_channel_views.xml
@@ -1,0 +1,145 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <!-- SLIDE.CATEGORY -->
+        <record model="ir.ui.view" id="view_slides_category_form">
+            <field name="name">slide.category.form</field>
+            <field name="model">slide.category</field>
+            <field name="arch" type="xml">
+                <form string="Category">
+                    <sheet>
+                        <group>
+                            <field name="name"/>
+                            <field name="channel_id"/>
+                        </group>
+                    </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record id="view_slides_category_tree" model="ir.ui.view">
+            <field name="name">slide.category.tree</field>
+            <field name="model">slide.category</field>
+            <field name="arch" type="xml">
+                <tree string="Category">
+                    <field name="sequence" widget="handle"/>
+                    <field name="name"/>
+                    <field name="channel_id"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="action_ir_slide_category" model="ir.actions.act_window">
+            <field name="name">Categories</field>
+            <field name="res_model">slide.category</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="help" type="html">
+                <p class="o_view_nocontent_smiling_face">
+                    Create a new category
+                </p>
+            </field>
+        </record>
+
+        <!-- SLIDE.CHANNEL VIEWS -->
+        <record model="ir.ui.view" id="view_slide_channel_form">
+            <field name="name">slide.channel.form</field>
+            <field name="model">slide.channel</field>
+            <field name="arch" type="xml">
+                <form string="Channels">
+                    <sheet>
+                        <div class="oe_button_box" name="button_box">
+                            <button class="oe_stat_button" name="website_publish_button"
+                                    type="object" icon="fa-globe">
+                                <field name="website_published" widget="website_button"/>
+                            </button>
+                            <button name="toggle_active" type="object" confirm="(Un)archiving a channel automatically (un)archives its slides. Do you want to proceed?" class="oe_stat_button" icon="fa-archive">
+                                <field name="active" widget="boolean_button" options='{"terminology": "archive"}'/>
+                            </button>
+                        </div>
+                        <div class="oe_title">
+                            <div class="oe_edit_only">
+                                <label for="name" string="Name"/>
+                            </div>
+                            <h1>
+                                <field name="name" default_focus="1" placeholder="Name"/>
+                            </h1>
+                        </div>
+                        <field name="category_ids" widget="many2many_tags" placeholder="Categories" context="{'default_channel_id': active_id}"
+                            domain="[('channel_id','=', active_id)]"/>
+                        <notebook colspan="4">
+                            <page string="General">
+                                <group>
+                                    <group>
+                                        <field name="website_id" options="{'no_create': True}" groups="website.group_multi_website"/>
+                                        <field name="promote_strategy" widget="radio"/>
+                                        <field name="custom_slide_id" class="oe_edit_only" attrs="{'invisible': [('promote_strategy', '!=', 'custom')]}" domain="[('channel_id','=',id), ('website_published','=',True)]"/>
+                                        <field name="promoted_slide_id" attrs="{'invisible': [('promote_strategy', '!=', 'custom')]}"/>
+                                    </group>
+                                    <group>
+                                        <field name="publish_template_id" domain="[('model','=','slide.slide')]"/>
+                                        <field name="share_template_id" domain="[('model','=','slide.slide')]"/>
+                                    </group>
+                                </group>
+                                <label for="description"/>
+                                <field name="description" colspan="4" nolabel="1"/>
+                            </page>
+                            <page string="Security">
+                                <group>
+                                    <field name="upload_group_ids" widget="many2many_tags"/>
+                                    <field name="visibility" string="Channel visibility" widget="radio"/>
+                                    <field name="group_ids" widget="many2many_tags" attrs="{'invisible':[('visibility', '=', 'public')], 'required':[('visibility', 'in',('private', 'partial'))]}"/>
+                                </group>
+                                <group attrs="{'invisible':[('visibility','in',('public'))]}" colspan="4">
+                                        <label for="access_error_msg" string="Error Message"/>
+                                        <field name="access_error_msg" colspan="4" nolabel="1"/>
+                                </group>
+                            </page>
+                        </notebook>
+                    </sheet>
+                    <div class="oe_chatter">
+                        <field name="message_follower_ids" widget="mail_followers"/>
+                        <field name="message_ids" widget="mail_thread"/>
+                    </div>
+                </form>
+            </field>
+        </record>
+
+        <record id="view_slide_channel_tree" model="ir.ui.view">
+            <field name="name">slide.channel.tree</field>
+            <field name="model">slide.channel</field>
+            <field name="arch" type="xml">
+                <tree string="Channel">
+                    <field name="sequence" widget="handle"/>
+                    <field name="name"/>
+                    <field name="website_id" groups="website.group_multi_website"/>
+                    <field name="active" invisible="1"/>
+                    <field name="visibility"/>
+                    <field name="promote_strategy"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="slide_channel_view_search" model="ir.ui.view">
+            <field name="name">slide.channel.search</field>
+            <field name="model">slide.channel</field>
+            <field name="arch" type="xml">
+                <search string="Channel">
+                    <filter string="Archived" name="inactive" domain="[('active','=',False)]"/>
+                </search>
+            </field>
+        </record>
+
+        <record id="action_slide_channels" model="ir.actions.act_window">
+            <field name="name">Channels</field>
+            <field name="res_model">slide.channel</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="help" type="html">
+                <p class="o_view_nocontent_smiling_face">
+                    Create a channel
+                </p>
+            </field>
+        </record>
+    </data>
+</odoo>
