+++ addons/website_slides/views/website_slides_templates.xml
@@ -2,22 +2,6 @@
 <odoo>
 <data>
 
-<!-- assets -->
-<template id="assets_frontend" inherit_id="website.assets_frontend" name="Slide">
-    <xpath expr="." position="inside">
-        <link rel="stylesheet" type="text/scss" href="/website_slides/static/src/scss/website_slides.scss" t-ignore="true"/>
-        <!-- Js widgets, ... order loading files is IMPORTANT -->
-        <script type="text/javascript" src="/website_slides/static/src/js/slides.js"/>
-        <script type="text/javascript" src="/website_slides/static/src/js/slides_upload.js"/>
-    </xpath>
-</template>
-
-<template id="assets_editor_inherit_website_slides" inherit_id="website.assets_editor" name="website_slides Assets Editor">
-    <xpath expr="." position="inside">
-        <script type="text/javascript" src="/website_slides/static/src/js/website_slides.editor.js"/>
-    </xpath>
-</template>
-
 <!-- Tools template: share on social networkds -->
 <template id='slides_share' name="Slides Media Share">
     <div>
